requirements.txt
streamlit
streamlit-option-menu
streamlit-folium
folium
Pillow